@@ -34,11 +34,11 @@
     "@types/node": "^17.0.4",
     "@types/react": "^17.0.38",
     "c8": "^7.10.0",
-    "eslint": "^8.6.0",
+    "eslint": "^8.7.0",
     "react": "^17.0.2",
-    "typescript": "4.6.0-dev.20220105",
-    "vite": "^2.7.7",
-    "vitest": "^0.0.115"
+    "typescript": "4.6.0-dev.20220114",
+    "vite": "^2.7.12",
+    "vitest": "^0.1.17"
   },
   "dependencies": {
     "@babel/core": "^7.16.5",
