@@ -27,15 +27,15 @@
     "@types/babel__core": "^7.1.17",
     "@types/jest": "^27.0.3",
     "@types/node": "^17.0.4",
-    "eslint": "^8.6.0",
-    "typescript": "4.6.0-dev.20220105"
+    "eslint": "^8.7.0",
+    "typescript": "4.6.0-dev.20220114"
   },
   "dependencies": {
     "@babel/core": "^7.16.5",
     "@reskript/config-babel": "3.0.2",
     "@reskript/core": "3.0.2",
     "cli-highlight": "^2.1.11",
-    "globby": "^12.0.2",
+    "globby": "^12.1.0",
     "p-limit": "^4.0.0"
   },
   "peerDependencies": {
