@@ -26,19 +26,19 @@
   "devDependencies": {
     "@types/jest": "^27.0.3",
     "@types/node": "^17.0.4",
-    "@types/ramda": "^0.27.62",
-    "typescript": "4.6.0-dev.20220105",
+    "@types/ramda": "^0.27.64",
+    "typescript": "4.6.0-dev.20220114",
     "webpack": "^5.65.0"
   },
   "dependencies": {
     "@reskript/config-webpack": "3.0.2",
     "@reskript/core": "3.0.2",
     "@reskript/settings": "3.0.2",
-    "eslint": "^8.6.0",
+    "eslint": "^8.7.0",
     "kolorist": "^1.5.1",
     "matcher": "^5.0.0",
     "pretty-bytes": "^5.6.0",
-    "ramda": "^0.27.1",
+    "ramda": "^0.28.0",
     "tty-table": "^4.1.5",
     "webpack-bundle-analyzer": "^4.5.0"
   },
