@@ -26,8 +26,8 @@
   "devDependencies": {
     "@types/jest": "^27.0.3",
     "@types/node": "^17.0.4",
-    "eslint": "^8.6.0",
-    "typescript": "4.6.0-dev.20220105",
+    "eslint": "^8.7.0",
+    "typescript": "4.6.0-dev.20220114",
     "webpack": "^5.65.0"
   },
   "dependencies": {
