@@ -24,20 +24,20 @@
     "test": "echo 'No test in @reskript/cli-lint'"
   },
   "devDependencies": {
-    "@types/eslint": "^7.29.0",
+    "@types/eslint": "^8.2.2",
     "@types/jest": "^27.0.3",
     "@types/node": "^17.0.4",
-    "@types/ramda": "^0.27.62",
-    "typescript": "4.6.0-dev.20220105"
+    "@types/ramda": "^0.27.64",
+    "typescript": "4.6.0-dev.20220114"
   },
   "dependencies": {
     "@reskript/config-lint": "3.0.2",
     "@reskript/core": "3.0.2",
-    "eslint": "^8.6.0",
+    "eslint": "^8.7.0",
     "eslint-formatter-pretty": "^4.1.0",
     "execa": "^6.0.0",
-    "globby": "^12.0.2",
-    "ramda": "^0.27.1",
+    "globby": "^12.1.0",
+    "ramda": "^0.28.0",
     "stylelint": "^14.2.0"
   },
   "peerDependencies": {
