@@ -27,20 +27,20 @@
     "@types/body-parser": "^1.19.2",
     "@types/debounce": "^1.2.1",
     "@types/dedent": "^0.7.0",
-    "@types/marked": "^3.0.3",
+    "@types/marked": "^4.0.1",
     "@types/mdast": "^3.0.10",
     "@types/node": "^17.0.4",
-    "@types/ramda": "^0.27.62",
+    "@types/ramda": "^0.27.64",
     "@types/react": "^17.0.38",
     "c8": "^7.10.0",
     "cpy-cli": "^3.1.1",
     "dedent": "^0.7.0",
-    "eslint": "^8.6.0",
+    "eslint": "^8.7.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
-    "typescript": "4.6.0-dev.20220105",
-    "vite": "^2.7.7",
-    "vitest": "^0.0.115",
+    "typescript": "4.6.0-dev.20220114",
+    "vite": "^2.7.12",
+    "vitest": "^0.1.17",
     "webpack": "^5.65.0"
   },
   "dependencies": {
@@ -54,14 +54,14 @@
     "debounce": "^1.2.1",
     "internal-ip": "^7.0.0",
     "localforage": "^1.10.0",
-    "marked": "^3.0.8",
-    "ramda": "^0.27.1",
-    "remark-gfm": "^1.0.0",
-    "remark-parse": "^9.0.0",
-    "remark-stringify": "^9.0.1",
+    "marked": "^4.0.10",
+    "ramda": "^0.28.0",
+    "remark-gfm": "^3.0.1",
+    "remark-parse": "^10.0.1",
+    "remark-stringify": "^10.0.2",
     "socket.io": "^4.4.0",
     "socket.io-client": "^4.4.0",
-    "unified": "^9.2.2",
+    "unified": "^10.1.1",
     "webpack-dev-server": "^4.7.1"
   },
   "peerDependencies": {
