@@ -26,16 +26,16 @@
   "devDependencies": {
     "@types/jest": "^27.0.3",
     "@types/node": "^17.0.4",
-    "@types/ramda": "^0.27.62",
-    "eslint": "^8.6.0",
-    "typescript": "4.6.0-dev.20220105"
+    "@types/ramda": "^0.27.64",
+    "eslint": "^8.7.0",
+    "typescript": "4.6.0-dev.20220114"
   },
   "dependencies": {
     "@reskript/config-jest": "3.0.2",
     "@reskript/core": "3.0.2",
     "@reskript/settings": "3.0.2",
     "jest-cli": "^27.4.5",
-    "ramda": "^0.27.1"
+    "ramda": "^0.28.0"
   },
   "peerDependencies": {
     "core-js": "3.x"
