@@ -36,8 +36,8 @@
     "@reskript/cli-test": "3.0.2",
     "@types/jest": "^27.0.3",
     "@types/node": "^17.0.4",
-    "eslint": "^8.6.0",
-    "typescript": "4.6.0-dev.20220105"
+    "eslint": "^8.7.0",
+    "typescript": "4.6.0-dev.20220114"
   },
   "dependencies": {
     "@reskript/core": "3.0.2",
