@@ -3,7 +3,7 @@ import {LoaderFactory} from '../interface.js';
 
 const factory: LoaderFactory = async () => {
     return {
-        loader: MiniCssExtractPlugin.loader,
+        loader: MiniCssExtractPlugin.default.loader,
     };
 };
 
