@@ -2,20 +2,23 @@ import path from 'path';
 import fs from 'fs';
 import webpack from 'webpack';
 import HTMLWebpackPlugin from 'html-webpack-plugin';
-import InteroplateHTMLPlugin from '../index';
+import {dirFromImportMeta} from '@reskript/core';
+import InteroplateHTMLPlugin from '../index.js';
+
+const currentDirectory = dirFromImportMeta(import.meta.url);
 
 export default (replacements: Record<string, string>) => {
     const compiler = webpack({
         devtool: false,
         mode: 'development',
-        context: __dirname,
+        context: currentDirectory,
         entry: './fixtures/index.js',
         output: {
-            path: path.join(__dirname, 'output'),
+            path: path.join(currentDirectory, 'output'),
             filename: 'bundle.js',
         },
         plugins: [
-            new HTMLWebpackPlugin({template: path.join(__dirname, 'fixtures', 'index.html')}),
+            new HTMLWebpackPlugin({template: path.join(currentDirectory, 'fixtures', 'index.html')}),
             new InteroplateHTMLPlugin(replacements),
         ],
     });
@@ -38,7 +41,7 @@ export default (replacements: Record<string, string>) => {
                 reject(new Error(result?.errors?.[0].message ?? 'Unknown error'));
             }
 
-            const output = fs.readFileSync(path.join(__dirname, 'output', 'index.html'), 'utf-8');
+            const output = fs.readFileSync(path.join(currentDirectory, 'output', 'index.html'), 'utf-8');
             resolve(output);
         });
     });
